tests separated and conftest added
